+++ tempodb/blocklist/list.go
@@ -119,7 +119,7 @@ func (l *List) Update(tenantID string, add []*backend.BlockMeta, remove []*backe
 }
 
 // updateInternal exists to do the work of applying updates to held PerTenant and PerTenantCompacted maps
-//  it must be called under lock
+// it must be called under lock
 func (l *List) updateInternal(tenantID string, add []*backend.BlockMeta, remove []*backend.BlockMeta, compactedAdd []*backend.CompactedBlockMeta, compactedRemove []*backend.CompactedBlockMeta) {
 	// ******** Regular blocks ********
 	blocklist := l.metas[tenantID]